ENH: Add select methods for IPC
Allows for select statements that returns memory handles, rather than a result
set. ATM this is implemented on the `connection` object, rather than the cursor.
I'm not sure which is more appropriate. For it to belong on the `Cursor` object,
I think we would need to bring the parsing code from pygdf into pymapd, so we
can parse the result into a result set, and return an Iterator over that.