Swap itertuples for .values
Calling .values on a dataframe creates an array, promoted to the common 
datatype able to hold all the data. This can cause ints to be converted 
to float and other undesirable behavior.

Intent is just to move from dataframe to an iterable, which itertuples 
does, without changing data representation in a cell