Throw exception if version number <4.6 (#220)
With #188, we need to prevent users using OmniSci <=4.5 from using pymapd >= 0.11. This is because of the backend change to switch encoding of dates from epoch seconds to epoch days. It was determined by myself and JP that it's better to throw the error on login, rather than allow user to have a script blow up unexpectedly once using columnar load.