Remove legacy typing (#222)
Per #161, type annotations are likely out-of-date. Currently, these annotations aren't being checked via CI, nor are they in the correct format for Python 3 (they are in Python 2 comment style).

Removing for now to provide a clean slate, so that if someone in the future chooses to incorporate typing, they will not have to wade through the legacy comments.